1.调整例子，打印 RT 时间。

+++ test-rsf-server/src/main/java/test/net/hasor/rsf/service/EchoService.java
@@ -20,5 +20,5 @@ package test.net.hasor.rsf.service;
  * @author 赵永春(zyc@hasor.net)
  */
 public interface EchoService {
-    public String echo(String sayMessage, Integer i);
+    public String echo(String sayMessage, long i);
 }
